+++ src/pyFAI/detectors/_common.py
@@ -733,10 +733,10 @@ class Detector(metaclass=DetectorMeta):
                 p1, p2, p3 = bilinear.calc_cartesian_positions(d1c.ravel(), d2c.ravel(),
                                                                self._pixel_corners,
                                                                is_flat=self.IS_FLAT)
-                p1.shape = d1.shape
-                p2.shape = d1.shape
+                p1 = p1.reshape(d1.shape)
+                p2 = p2.reshape(d1.shape)
                 if p3 is not None:
-                    p3.shape = d1.shape
+                    p3 = p3.reshape(d1.shape)
             else:
                 i1 = d1.astype(int).clip(0, self._pixel_corners.shape[0] - 1)
                 i2 = d2.astype(int).clip(0, self._pixel_corners.shape[1] - 1)