+++ src/pyFAI/test/test_blob_detection.py
@@ -4,7 +4,7 @@
 #    Project: Azimuthal integration
 #             https://github.com/silx-kit/pyFAI
 #
-#    Copyright (C) 2015-2025 European Synchrotron Radiation Facility, Grenoble, France
+#    Copyright (C) 2015-2026 European Synchrotron Radiation Facility, Grenoble, France
 #
 #    Principal author:       Jérôme Kieffer (Jerome.Kieffer@ESRF.eu)
 #
@@ -106,7 +106,13 @@ class TestMorphology(unittest.TestCase):
         res_d = morphology.binary_dilation(self.img)
         self.assertTrue(numpy.allclose(ref_d, res_d), "binary dilation")
         res_e = morphology.binary_erosion(res_d)
+        res_e = morphology.binary_erosion(res_d)
         self.assertTrue(numpy.allclose(self.img, res_e), "binary erosion")
+
+    def test_peak_search_before_process(self):
+        bd = BlobDetection(self.img)
+        with self.assertRaisesRegex(RuntimeError, r"No keypoints yet: run BlobDetection\.process before peak search"):
+            bd.peaks_from_area(numpy.ones_like(self.img, dtype=bool))
 
 
 def suite():