+++ src/pyFAI/test/test_goniometer.py
@@ -4,7 +4,7 @@
 #    Project: Azimuthal integration
 #             https://github.com/silx-kit/pyFAI
 #
-#    Copyright (C) 2017-2025 European Synchrotron Radiation Facility, Grenoble, France
+#    Copyright (C) 2017-2026 European Synchrotron Radiation Facility, Grenoble, France
 #
 #    Principal author:       Jérôme Kieffer (Jerome.Kieffer@ESRF.eu)
 #
@@ -33,7 +33,7 @@ __author__ = "Jérôme Kieffer"
 __contact__ = "Jérôme.Kieffer@esrf.fr"
 __license__ = "MIT"
 __copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "24/02/2026"
+__date__ = "05/06/2026"
 
 import os
 import math
@@ -43,7 +43,7 @@ from .utilstest import UtilsTest
 import numpy
 from ..goniometer import GeometryTranslation, Goniometer, numexpr, \
                          ExtendedTransformation, GoniometerRefinement,\
-                         PoniParam
+                         PoniParam, SingleGeometry
 logger = logging.getLogger(__name__)
 
 
@@ -190,6 +190,29 @@ class TestTranslation(unittest.TestCase):
         self.assertAlmostEqual(param, gonio.trans_function.ParamNT(1, 2, 3, -4, 6, math.pi/2, 1.7712028347600038e-10))
         if os.path.exists(fname):
             os.unlink(fname)
+
+    def test_extract_points_no_image(self):
+        gonio = SingleGeometry(
+          label="test_cps",
+          detector="pilatus100k",
+          image=None,
+          geometry=None,
+          calibrant=None,
+        )
+        with self.assertRaisesRegex(RuntimeError, "To perform control point extraction, a data image must be provided"):
+          gonio.extract_cp()
+          
+    def test_extract_points_no_wavelength(self):
+        gonio = SingleGeometry(
+          label="test_cps",
+          detector="pilatus100k",
+          image=numpy.zeros((195, 487)),
+          geometry=None,
+          calibrant=None,
+        )
+        with self.assertRaisesRegex(RuntimeError, "To perform control point extraction, a wavelength must be provided"):
+          gonio.extract_cp()
+
 
 def suite():
     loader = unittest.defaultTestLoader.loadTestsFromTestCase